fixed spiraling

+++ sgl/src/sgl_core.cpp
@@ -169,5 +169,5 @@ void SglCore::draw_arc(SglVertex center, float radius, float from, float to) {
 
     radius = radius * std::sqrt(get_scaling_factor());
 
-    renderer.draw_arc(center, static_cast<int>(radius), from, to, contexts.at(current_context).matrix_stacks[sglEMatrixMode::SGL_MODELVIEW].top());
+    renderer.draw_arc(center, radius, from, to, contexts.at(current_context).matrix_stacks[sglEMatrixMode::SGL_MODELVIEW].top());
 }
